requirements.txt
langchain==0.3.27
langchain-core==0.3.75
langchain-community==0.3.29
langchain-groq==0.3.5
langchain-google-genai==2.1.8
langgraph==0.6.6
pydantic==2.11.7
mysql-connector-python==9.3.0
pyodbc==5.2.0
psycopg2==2.9.10